requirements.txt
pandas==1.5.3
python-dotenv==1.0.0
pint==0.20.1
datetime==5.0
sqlalchemy==1.4.40
pyarrow==11.0.0